Correct parameters for generate-constraints
When proposing updates, the parameter passed to the update function
is the name of the subdirectory where the project to be updated can
be found, so the global-requirements.txt is relative to the
requirements repo workspace. Also when passing it to the
generate-constraints entrypoint, do so with the -r parameter rather
than as a positional argument (since the latter does not work).

Change-Id: I22acc4598dda92a60b47d34a7986a117197f8626