Added eslint-config-openstack
This project is intended to be a javascript equivalent of openstack's
"hacking" project. It will also act as a testbed for any openstack tools
used to publish this project to the npm repository.

Rules within the project have been set to what is currently under
discussion in the 'depends-on' patch. The openstack namespace has been
chosen over openstack-dev, as I was informed that the latter is being
deprecated in favor of the former.

Channel notifications have been added to openstack-javascript,
openstack-merges, and openstack-qa. Projects that use these linting
rules should add their own channel subscriptions.

Change-Id: I434465d126ee22bd7fb5e1f9d6d77763640aa3d9
Depends-on: Ib4c3f77f8cc3cdaa3c7558b7bc3a6d1299b6dcbe