Do not set ZUUL_REF to a valid value on devstack checkout
Zuul mergers do not keep all repository-branches up to date all of
the time.  Instead, they are current only under two conditions:

1) if the zuul ref is refs/zuul/.* and they have that ref
2) if the zuul ref is anything and the project matches zuul_project

Condition 1 is true during tests on changes, condition 2 is true
for tags and timers.

However, this job runs on the "tempest" project, meaning that
neither of the conditions above hold, and therefore zuul mergers
should not be used to check out the repositories.  However, because
we gave zuul-cloner a ref that actually is in a merger repo
("master"), it behaved as if condition 1 were true and checked it
out.  Instead, give it an invalid ref so that it falls back to the
canonical git repository to get the correct master branch-tip.

Change-Id: I0c2402979432e86c7a3ba4ba705486db74ea4db9