puppet/upgrade: do not reinstall Puppet
In Trusty: beaker is running dpkg with '--force' option so even if
Puppet is already installed, the return code will be 0.
In CentOS: beaker is running 'rpm -ivh' the second time and since the
package is already here, the return code will be 1.

Beaker does not need to be re-install Puppet during the second
provisioning during the upgrade flow. So let's define 'BEAKER_provision'
to 'no' so we don't try to install Puppet a second time.

Change-Id: I1834269ee852cc9e7b66f00622e041cafaee2251