Fix Normalize projects.yaml proposal
The proposals bot run on a tree that is in detached state. Current
git review fails to push from this any changes.

Create instead directly a branch and work on that.

This is the same change as done for other proposal jobs in
I487337cb1def5f5995d776c831ea61b658652677.

Change-Id: I3ca3cbb0c21efde28a43164c518405c0dea87a86