Experiment with an increase time for gate-trove-functional job
Recently there have been many failures of this job at the current 105m
timeout. However, in some cases, it has been observed that the job is
being failed at 105m.

There appears to be considerable variability in the performance of the
hypervisors used for the CI and I would like to increase the time
(considerably) as proposed in this change and then measure the actual
time it takes to pass consistenly and then resubmit this change with a
more reasonable increase over 105m.

I could increase this by 5minutes and see if it worked or not, and
then come back if it didn't but it is generally the case that a binary
search will converge more quickly on a solution.

Hence, I'm proposing 145 minutes as a starting point in this search.

Change-Id: I7478dde226f5ea2d6b0777e234b53e66ebcb8e57