Fix project requirements check to work with "extras"
Process "extras" defined in a project setup.cfg properly by looping over
the dictionary's member items and not the dictionary itself.

Do not compare comments associated with requirements, since those are
always collected as a set of unnamed requirements and many of the
comments from global-requirements.txt are not needed in the local
project file (especially when the dependencies for which the comments
are relevant are not used).

Change-Id: I78838dcd4da43b3c1d2610ac87a3ec55b9535646