Directly edit the unmapped locale from zanata.xml
We map all locales that contain a _ so that they map correctly to the
locale present on the Zanata server. While testing the pull
functionality, it turns out the client throws a traceback if there are
duplicate locales, so just edit the existing locale element, rather
than adding potential duplicates.

Change-Id: I9fc8df1996a5229bfef8e35f6d2fe62b06f6d01a