Create clone map outside workspace in zuul-.*-prep
Since zuul-cloner needs a clone map to tell it to clone the tested
project as the job workspace, and git won't clone into a non-empty
directory, the clone map needs to be written outside the workspace.
Use a temporary file, and clean it up with an exit trap in case this
macro is used on long-lived job workers.

Change-Id: I40ebd3dbb9123edce973d8ddaa06f0f804e16c56