Make Rally jobs for Manila non-votable
Recently, Rally jobs for Manila 'gate-rally-dsvm-manila-multibackend-no-ss' and
'gate-rally-dsvm-manila-multibackend' were added to check queue to run on each
change. Make these jobs as non-votable as they are used to indicate that Rally
code for Manila works, but it is not desired to block Rally changes because of
possible temporary breakages of Manila project.

Change-Id: I7f68c6bf56d11e113262c880be368a532624b72d