Remove neutron smoke job from integrated gate templates
This commit removes the neutron smoke job from the integrated gate
templates. This job is the same as the neutron full except only runs
a very small subset of the tempest tests instead of all of tempest
like the full job. This is just wasting a node for every commit
because it doesn't increase our coverage over the other job. On
the neutron integrated gate template a grenade job which was missing
from the gate queue was also added since it was missing.

Change-Id: I278361067ca058df953fc0f15db0261cab996891