oslo.db: disable python34 job for stable/juno
Now when python34 job is voting by default, stable/juno in oslo.db
is blocked, as test-requirements.txt contains MySQL-Python, which
fails to install on Python 3.x.

In master we switched to PyMySQL as a default MySQL driver and
stable/kilo has separate lists of test-requirements for Python 2.x
and 3.x respectively, so we only need to disable the job for
stable/juno.

Change-Id: If454ab6cfcd68f7ea69794652ff5359270c00776