puppet: introduce puppet-openstack-integration job
New job: puppet-openstack-integration
Project: openstack/puppet-openstack-integration (for now, will gate
other Puppet modules later)
Voting: false (for now, will change later).
Branch: master (for now, can change later).
Systems: centos7 and trusty

This job aims to gate Puppet OpenStack modules by deploying OpenStack
with Puppet and testing with Tempest so we can validate the full
deployment workflow, and not only one module like it's done with beaker
jobs.

For now, this patch only introduces the basic structure to run the
`run_tests.sh` script that will execute the provisionning.

It's non-voting because it's experimental.
It's running only on master patches because we don't have integration
tests on stable branches.
It's running only on openstack/puppet-openstack-integration because we
first want to bootstrap the code and have the structure working, then we
will add the jobs for other projects (Puppet OpenStack modules).

Change-Id: I71be8a26b30389ae13652ad7d333fcb2173a27e7