neutron: fix extension_drivers setting for feature/qos
It turns out that devstack does not apply post-config sections from
localrc but local.conf only [1]. Since devstack-gate uses the former to
configure setup, the hook was not applied.

ml2 plugin for devstack actually allows to pass the desired extensions
thru Q_ML2_PLUGIN_EXT_DRIVERS= variable, so let's use it instead of
making devstack post-config hooks localrc-aware.

This is already the third patch we have to enable the extensions in
gate. Sigh.

[1]: http://git.openstack.org/cgit/openstack-dev/devstack/tree/stack.sh#n1152

Change-Id: I51f28b63400b6322baed7ac28715590c391ccebd