Move puppet-syntax gate to devstack-centos7
The main reason is to move away from the bare nodes, since -infra will
be retiring in the future.  Additionally, we move into centos7 to
avoid the hurt of other projects using devstack-trusty.

Since we are installing puppet from gems, we don't need to rely on the
OS as much. However, if we want to add a devstack-trusty node in the
future, we now can easily.

Change-Id: I8d19231a5eccbfec7c0e035077aab47bf4ecc1b1
Signed-off-by: Paul Belanger <pabelanger@redhat.com>