Move puppet-unit gate to devstack-centos7
The main reason is to move away from the bare nodes, since -infra will
be retiring in the future.  Additionally, we move into centos7 to
avoid the hurt of other projects using devstack-trusty.

Since we are installing puppet from gems, we don't need to rely on the
OS as much. However, if we want to add a devstack-trusty node in the
future, we now can easily.

Change-Id: I79d216c4c84ebc954caef627ce64d7e72fdad8ff
Signed-off-by: Paul Belanger <pabelanger@redhat.com>