Create experimental jobs for VPN DevStack plugin
Now that VPN has a devstack plugin, we need to make sure that it is enabled
for the python-neutronclient. Will do this in three steps, so that the
migration to the plugin use can be done without affecting working jobs.

First, in Ieeaca1375d68705509f4e05f10cb35c0fa0b9582, the python-neutronclient
code added a gate hook and changes to be able to support new jobs.

Second, in this commit, we'll add experimental jobs that create a gate hook
so that the needed customization can be done. There will be two jobs, one for
core tests (neutronclient-test-dsvm-functional), and one for advanced services
tests (neutronclient-test-dsvm-functional-adv-svcs).

Third, I1e3d19e51a1cbd1bc947bbf9927260cd4d73841a will use the new hook, enable
the VPN devstack plugin for the new experimental advanced services based job,
split the VPN tests out into a separate directory to run under the advanced
services job, and will move the rest of the tests to a "core" subdirectory.
The new experimental job will be tested by this commit.

Once tested, the experimental jobs can be made voting, and the old job
removed.

Change-Id: I2db939bf99288c0cdec06cdd49fec3bdc72e5253
Partial-Bug: 1484148
Depends-On: Ieeaca1375d68705509f4e05f10cb35c0fa0b9582