Fix coverage report links on logs.o.o
If we copy the console.log but not cover report to the logs server then
copy console log and cover report to swift os-loganalyze and apache will
serve up the mod autoindex generated index that only shows the console
log and does not show the cover report. This is because the cover report
is only on swift and not on disk.

We can ensure that the cover report always shows up in the index by not
copying any logs on disk and using swift exclusively.

Change-Id: I2ebcc33f8dcb03eb3674ff702f270d8cdca0afde