Skip non-doc and test-only jobs for for all branches
There is no much reason to skip it for master but not feature and
stable branches. And it only makes life harder for those stuck with
their features in feature branches.

While at it, expanded the filter to neutron-fullstack job to skip unit
test changes and other changes that are irrelevant for gate without
devstack.

Finally, moved other dsvm filters under more wide filter common to all
projects.

Change-Id: I4954a2aa7b2ede85843ccd51686bf648dd3782c6