zuul/layout: extend puppet-openstack-integration testing
To avoid any future issue in puppet-* modules when patching
puppet-openstack-integration, run
gate-puppet-keystone-puppet-beaker-rspec-dsvm-centos7 job.

Why:
changes in the integration repo affect the gates of the other modules,
so we want to run some jobs from one of those modules as a "canary"
to try to protect against accidentally breaking the gates of 25 other
repos.

Note:
* puppet-keystone because this is our most popular module and usually
  gets general changes the first, before other modules.
* centos7 because dsvm jobs are scheduled & run faster on this plateform.

Change-Id: I120463f960918588ea62b507afa9b4acbe508fe3