Add docs-on-rtfd to auto-publish docs
Right now, the os-ansible-deployment project doesn't have automatically
published docs, and we're waiting on a project rename and move before
pushing to the openstack.org docs location.

Primarily, we have submitted for the project to go from
'stackforge/os-ansible-deployment' to 'openstack/openstack-ansible'.
Since this is not just a namespace move but also a project rename, we're
holding off on a docs.openstack.org job to reduce any potential
conflict. See https://review.openstack.org/#/c/191105/ for more
information.

This change allows us to automatically build to readthedocs.org in the
meantime.

Change-Id: I9df80688ac18385c0fcef7628a0e8fc7265545c8