Added npm-publish-tarball
This adds an npm job that runs npm-pack, and then renames the
tarballs to match our versioning scheme before uploading them to
the tarball host.

The `npm pack` command commits the current project to the local
npm cache directory, which triggers a full lifecycle run (prepublish,
etc). This generates a tarball with all assets listed in the package's
"files" field, inside the npm cache. This tarball is then copied back
to the working directory.

Since the default naming is not helpful, the job then immediately
renames it before the tarballs are published, using our usual
version names.

This job is then used by ironic-webclient. Note that this commit
removes 'bare-precise' as a node option, because Ubuntu precise does
not package a recent enough version of nodejs for any of this.

Change-Id: Ifb7898f136fafae70b5b0cd2727cc478034667c7