Add experimental job for Barbican to test updated gate
We're working on updating the Barbican devstack plugin, and some changes
would be breaking for the existing gate. This experimental job is to
allow us to test the new job definition without risking temporary
breakage to the gate. This is the change on the Barbican side:

https://review.openstack.org/#/c/167885/

This experimental job is a very minor tweak of the existing gate job, so
hopefully merging this is trivial. It will be removed as soon as it is
shown to work and the primary gate job is updated to match.

Change-Id: I3ceea997a391c8c3fc8b416f6c2d3bb876315cf3