Add `npm shrinkwrap` to `npm run` macro.
The npm shrinkwrap command is equivalent to pip freeze,
as it captures current dependency versions and generates a
deterministic list in npm-shrinkwrap.json. This file is then output
to the console for informational purposes, however unless the
project explicitly includes it in package.json's "files" field,
it will not be included in the release package.

Change-Id: I17ad8653dfed873092b811c41c638ff0ed0eb45d