kolla: Remove kolla-functional-f21 job
Remove the kolla-functional-f21 job because it only runs the
removed-from-kolla bash build scripts.  The build gating is now
handled by other jobs which run a Python based build tool.  As a
result, evreything related to the kolla-functoinal-f21 job can be
removed.

As stands now this check will always fail because the bash buid
scripts are completely removed  from the kolla repo and this check
job depends on these removed bash build scripts.

Change-Id: I4de8fb52c6a32b6999f6be4112fa0b76e9f0ac42