Promote horizon selenium test job
This moves the Selenium test job for Horizon out of
experimental into check, but makes it non-voting so
we can see what the broader impact of it is on the
patches. There has historically been some instability
and we hope to resolve this.

The experimental job was tested in these patches:
https://review.openstack.org/#/c/215408/
https://review.openstack.org/#/c/215435/
https://review.openstack.org/#/c/156572/

(we are in the process of getting the selenium suite
to pass, hence those all result in failures)

Change-Id: I7069581c084156a2953fb492a89daaa95620efc3