copy_puppet_logs: copy tempest.conf in txt file
To be able to browse the file from a Web Browser without downloading the
file on the local system, rename tempest.conf into tempest.conf.txt so
it will be compressed to be at the end tempest.conf.txt.gz.
It will still take less space but also will be readable easily from the
Web Browser.

Change-Id: I0d8d016d3cb1a328c2852cb746b3d33f6ae332a9