zuul/layout: make non-voting some trusty beaker jobs
Ubuntu packaging for Liberty - called Liberty Staging - is broken for
one month for heat, ironic and python-glanceclient.
It currently blocks our CI to land patches.

To unblock the situation, this patch aims to make the 3 jobs non-voting
but still running so we can monitor them and revert this patch when
packages are stable enough.

Jobs that we want non-voting:
* gate-puppet-glance-puppet-beaker-rspec-dsvm-trusty
* gate-puppet-heat-puppet-beaker-rspec-dsvm-trusty
* gate-puppet-ironic-puppet-beaker-rspec-dsvm-trusty

Change-Id: Id34550890f3dabb1cc3fd30dda6583b5847f0cd2