make gate-tempest-dsvm-ironic-pxe_ssh non voting
Currently this is wedging all number of things because ironic is using
diskimage builder to pull things from ubuntu upstream and a name
change in their images meant things no longer SHA256 correctly.

We should make this non-voting to fix things, and ensure we're not
making it voting again until this is no longer pulling from the open
internet.

Change-Id: I9002e32126a3f07efa857aa60f5431b409edb8b9