zuul/layout: make non-voting puppet-keystone beaker on trusty
Ubuntu Liberty packaging (still considered as unstable) is currently
broken for keystone package.

Since keystone is the common project used by any other modules, we can't
get stuck until they fix the package.
The pacakge works for centos7 so we still have beaker jobs running on
RDO trunk, so in the meantime Ubuntu fix the package, let's set it as
non-voting so it will unblock our CI.

This patch is temporary and will probably be reverted as soon Ubuntu
provides stable enough packages for Keystone and other broken services.

Change-Id: I6821dab1f204b83cc0d0d548de8de602aeaf8a6b