puppet-openstack/zuul-layout: make voting more ubuntu beaker jobs
Canonical released today OpenStack 'b2' which is packaged in 'proposed'
UCA repository.
The repository is stable enough to consume it from Puppet OpenStack CI
instead of painful Liberty Staging repo which is broken all the time.

This patch aims to re-enabling voting for all beaker jobs running on
Ubuntu except for Ironic where there is still issues with the package.
In the meantime it's being fixed, let's make it non-voting.

Change-Id: I08bb30bbd71ac083d24ca30607fc7e319dee21eb