Add python jobs to the openstack-health repo
Python code was recently added to the openstack-health project. This
is primarily to implement a rest api for the client side JS to leverage
when running the dashboard. However now that there is python code in the
repo we need to be able to run python tests as part of gating changes,
this commit adds the necessary pieces.

Change-Id: I3492eda5c88d4d5465924c326b195349f0a9bd87