Don't run python jobs for old Fuel Client
Support for python jobs was only added in the current version
which is 7.0. Therefore it's necessary to not block any patch
for any of 6.1.* versions that does not pass python tests on
OpenStack CI. Since Python-3 support is currently being
implemented, it's necessary to not run that job on 7.0 stable
branch.

Change-Id: I547145cd42001f349cefeac284d6b9e83b15bae1
Closes-bug: #1495890