No longer read an existing zanata.xml
There is code in ZanataUtils that will read an existing zanata.xml,
and parse the details into an instance to support changing an
existing XML file. We have never made use of this functionality,
and it just tripped me up when testing a change for the last time,
destroy it.

Change-Id: Ie3e3ba3d8171ae912f2706dfeb756c007b51b703