Updated npm-run macro to support npm 1.x
npm@1 returns an exit code 0 if an invalid command is passed. Later
versions (2+) return exit code 1. This makes bash -xe invocations
as used in jjb unreliable, as an "failed" invocation still returns
exit code 0.

This patch fixes this by explicitly enumerating valid npm lifecycle
phases. The command passed is checked against that list, and if
found will treat the command as an `npm {command}` event. If the
command is not found in this list, it is instead executed as
`npm run {command}`. By performing an explicit check, we no longer
have to rely on unreliable exit codes.

Change-Id: I38b0d066626a27d6be897b96c76c192157ee1463