Remove networking-midonet tempest-dsvm jobs
These jobs clone and build a repository from github within a
devstack plugin.  In order to do so, they must disable ERROR_ON_CLONE
which is our assurance that all tested OpenStack projects are in
the $PROJECTS variable.

Further, OpenStack projects must not clone non-OpenStack repositories
in gate tests -- they should either be consuming OpenStack projects
from git via the $PROJECTS variable or installing released packages
as dependencies.  For Midonet, that likely either means via the
operating system, or via a maven repository.  If via maven, then we
will need some commitment from the Midonet team to set up a maven
mirror similar to our pypi mirror so that when we run tests we are
not relying on external network resources.

Change-Id: I9194069c4754439977a11404bbc5ac650c5134df