Create experimental grafyaml gate job
Because we are planning on including grafyaml dashboards in
project-config we'll need to properly gate on them.  A follow up patch
will be adding tox -egrafyaml to do just this. Additionally, this job
will only match with tox.ini and the contents of grafana change.

Additionally, the grafyaml dashboards created in the follow up patch
are not actually consumed into production yet.  We still need to add
the puppet logic to handle that.

Change-Id: I497801f2de6dbfcb2b9644bedf361b5fa66463fd
Signed-off-by: Paul Belanger <pabelanger@redhat.com>