Create bareon project
This project is to provide a flexible and easy to
use operating system installation tool. It is to focus
on providing advanced partitioning functionality as well
as advanced pre-configuration features. It differs from
existent great cloud oriented solutions like Ironic Python
Agent (OS provisioning) and Cloud init (OS pre-configuration).
It is to be a tool similar to old plain OS installation tools like
anaconda or debian-installer but dramatically more convenient.

Change-Id: Ia6c59aefb260712d34d73a49f186e072637d7de5