Prevent errors due to non-existing console.html
Grep and stat were producing the error "No such file or directory"
while the consoleLog was not available to download. Keeping an empty
/tmp/console.html resolves the problem. It's also possible to debug any
possible issues with curl and grep independently since curl would
report 404 for missing url and a new error message was introduced
for missing UUID after the max retries.

Change-Id: Icb28bfca96e5eaff725d2efcfb47c3399e3991c4