Add javascript jobs to openstack-health
This commit adds running javascript jobs to the openstack-health
project. The openstack-health repo contains 2 projects in 1 a REST
API on top of subunit2sql (and eventually other data sources) to query
for results history and a javascript frontend to visualize the data
returned over that rest api. Currently the repo only runs tests on
python which only covers the REST API. This commit adds testing to
the other half of the repo on the JS side.

Change-Id: Ie467df9c1f9333c04a75ca62daad82e86768f254