Make javascript jobs on openstack-health non-voting
These jobs don't even come close to working. npm is going out to the
internet to pull packages and it is failing to download packages. This
is causing everything to fail before it even gets started. This commit
makes these jobs non-voting, it would have been good to have voting
coverage on the js in the repo, but if the jobs can't reliable pull in
requirements we can only use them as advisory jobs.

Change-Id: I0800b132cd792da0450cee620b953ee93b75be0a