Be explicit about projects list for reqs and pbr
The requirements and pbr jobs currently rely on devstack-gate to have an
opinion on which projects are interesting. Unfortunately this means all
d-g jobs ever setup all the projects which takes a lot of time. We can
save time by removing opinion from d-g and pushing it into the jobs
themselves. This change ensures that we retain the same level of test
coverage with requirements and pbr by pushing the old projects list into
their integration jobs.

Depends-On: I875641a5d00bc8afad97cbca6760ce66ac4a5c15
Change-Id: Ife6161a4b3cfce5041fe8e6f2c128dee14208b8c