Add a ceilometer-functional job for mysql
Add a ceilometer-functional job for mysql to run alongside the
postgresql and mongodb jobs, initially as an experimental job.

This job was probably supposed to be there all along, its absence an
oversight. Once it is confirmed to be happy a followup patch will
put it in gate and check.

Change-Id: I334db9864e1295b968847f646966b27556949d7b