Ability to test oslo master libraries against liberty services
During the Tokyo summit discussions, there were concerns raised
about backward compatability guarantees with oslo libraries. The
main reason was that in liberty global-requirements we do not
cap the versions for oslo libraries and any oslo release in
support of mitaka is likely to break the stable/liberty branch.
We need to add additional jobs to make sure a change in oslo
library in master does not break liberty.

We had this support earlier which was removed in change id:
Ia40d8cbf8cef803617509aeca20a9f1ec79d0880

We cannot just revert that change since a lot of changes have
happened since April. We should run this new job against master
as a check job and also should not vote until next summit.

Change-Id: Idcbb3e8c71b0be97c7ab78c426ac363388f46633