macros/puppet-prepare-node: drop rdo-release
When running Puppet jobs, RDO repository is managed by Puppet.
If rdo-release package is installed, we will always install Liberty
(latest rdo-release package) and we don't want that.
Puppet CI is currently deploying Liberty in stable/kilo jobs, because of
this package.

We need to drop this package before any Puppet related job, so we make
sure Puppet manifests will manage the repositories that we actually want.

Change-Id: I3173e850dfdf43b868e4a180083c19c3cbe30ff3